create separate playground
