+++ sandbox/.gitignore
@@ -14,6 +14,8 @@
 !driver-adapters/**
 !d1
 !d1/**
+!query-compiler
+!query-compiler/**
 
 # Ignore node_modules
 **/node_modules
