+++ sandbox/query-compiler/.gitignore
@@ -1,0 +1,1 @@
+dev.db
