+++ sandbox/query-compiler/package.json
@@ -1,6 +1,6 @@
 {
   "private": true,
-  "name": "basic-sqlite",
+  "name": "query-compiler",
   "description": "Prisma development playground",
   "main": "index.ts",
   "scripts": {
@@ -14,12 +14,16 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@prisma/client": "../../packages/client"
+    "@prisma/client": "../../packages/client",
+    "@prisma/adapter-pg": "../../packages/adapter-pg",
+    "pg": "../../packages/adapter-pg/node_modules/pg"
   },
   "devDependencies": {
     "@types/node": "18.18.9",
+    "@types/pg": "../../packages/adapter-pg/node_modules/@types/pg",
     "prisma": "../../packages/cli",
-    "ts-node": "10.9.1",
-    "typescript": "5.2.2"
+    "ts-node": "*",
+    "tsx": "*",
+    "typescript": "*"
   }
 }
