+++ sandbox/query-compiler/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": [
+    "./**/*"
+  ],
+  "exclude": ["node_modules"]
+}
